+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+### v1.2.0 (11-Apr-2022)
+
+- Add `IConfigureBootstrapperOptions.ConfigureBootstrapperConfiguration(...)`.
+- Depreciated `IConfigureBootstrapperOptions.UseAdditionalBootstrapperConfigurationFiles(...)`.
+- `SerilogLoggerConfigurationExtensions` improvements.
+- Fixed typos.
+
 ### v1.1.0 (21-Feb-2022)
 
 - Add new public extension `SerilogLoggerConfigurationExtensions.ConfigureWithFailSafeDefaults(...)`.
-- Depreciated `SerilogConfigurationExtensions.UseLogging`.
+- Depreciated `SerilogConfigurationExtensions.UseLogging(...)`.
 - Updated dependencies.
 
 ### v1.0.0 (09-Dec-2021)